+++ guillotina_elasticsearch/commands/vacuum.py
@@ -21,12 +21,6 @@ import json
 import logging
 
 
-try:
-    from guillotina.utils import clear_conn_statement_cache
-except ImportError:
-    def clear_conn_statement_cache(conn):
-        pass
-
 logger = logging.getLogger('guillotina_elasticsearch_vacuum')
 
 GET_CONTAINERS = 'select zoid from {objects_table} where parent_id = $1'
@@ -132,8 +126,9 @@ class Vacuum:
                 while len(results) > 0:
                     records = []
                     for record in results:
-                        if record['zoid'] in (ROOT_ID, TRASHED_ID,
-                                              self.container._p_oid):
+                        if record['zoid'] in (
+                                ROOT_ID, TRASHED_ID,
+                                self.container._p_oid):
                             continue
                         records.append(record)
                         self.last_tid = record['tid']
@@ -149,7 +144,8 @@ class Vacuum:
                 new_oids = []
                 while (pos * PAGE_SIZE) < len(oids):
                     async with conn.transaction():
-                        cur = await conn.cursor(sql, oids[pos:pos + PAGE_SIZE])
+                        cur = await conn.cursor(
+                            sql, oids[pos:pos + PAGE_SIZE])
                         pos += PAGE_SIZE
                         page = await cur.fetch(PAGE_SIZE)
                         while page:
@@ -168,7 +164,6 @@ class Vacuum:
             from guillotina.db.transaction import HARD_CACHE  # noqa
             result = HARD_CACHE.get(oid, None)
         if result is None:
-            clear_conn_statement_cache(await self.txn.get_connection())
             result = await self.txn._cache.get(oid=oid)
 
         if result is None:
@@ -212,7 +207,8 @@ class Vacuum:
         try:
             conn = await self.txn.get_connection()
             sql = self.get_sql(CREATE_INDEX)
-            await conn.execute(sql)
+            async with self.txn._lock:
+                await conn.execute(sql)
         except Exception:
             pass
 
@@ -228,7 +224,6 @@ class Vacuum:
         checked = 0
         async for es_batch, index_name in self.iter_batched_es_keys():
             checked += len(es_batch)
-            clear_conn_statement_cache(conn)
             async with self.txn._lock:
                 sql = self.get_sql(SELECT_BY_KEYS)
                 records = await conn.fetch(sql, es_batch)
@@ -269,7 +264,7 @@ class Vacuum:
                             'Could not parse delete by query response. '
                             'Vacuuming might not be working')
 
-    def get_indexes_for_tids(self, tids):
+    def get_indexes_for_oids(self, oids):
         '''
         is there something clever here to do this faster
         than iterating over all the data?
@@ -279,8 +274,8 @@ class Vacuum:
             # check if tid inside sub index...
             prefix = index['oid'].rsplit('|', 1)[0]
             if prefix:
-                for tid in tids:
-                    if tid.startswith(prefix):
+                for oid in oids:
+                    if oid.startswith(prefix):
                         indexes.append(index['index'])
                         break
         return indexes
@@ -293,7 +288,9 @@ class Vacuum:
         })
         conn = await self.txn.get_connection()
         sql = self.get_sql(GET_CONTAINERS)
-        containers = await conn.fetch(sql, ROOT_ID)
+        async with self.txn._lock:
+            containers = await conn.fetch(sql, ROOT_ID)
+
         if len(containers) > 1:
             # more than 1 container, we can't optimize by querying by tids
             self.use_tid_query = False
@@ -301,7 +298,7 @@ class Vacuum:
         checked = 0
         async for batch in self.iter_paged_db_keys([self.container._p_oid]):
             oids = [r['zoid'] for r in batch]
-            indexes = self.get_indexes_for_tids(oids)
+            indexes = self.get_indexes_for_oids(oids)
             results = await self.utility.conn.search(
                 ','.join(indexes), body={
                     'query': {